Fix integer overflow in MPEG4Extractor
Fix interger overflow due to addition of uint32_t and int32_t.

Bug: 241354088
Test: Extracted the clip shared in bug by using mpeg4 extractor

+++ media/module/extractors/mp4/MPEG4Extractor.cpp
@@ -6712,7 +6712,7 @@ media_status_t MPEG4Source::fragmentedRead(
         const Sample *smpl = &mCurrentSamples[mCurrentSampleIndex];
         offset = smpl->offset;
         size = smpl->size;
-        cts = mCurrentTime + smpl->compositionOffset;
+        cts = (int64_t)mCurrentTime + (int64_t)smpl->compositionOffset;
 
         if (mElstInitialEmptyEditTicks > 0) {
             cts += mElstInitialEmptyEditTicks;
